Move Mustache patch to relevant code that uses it
Side benefit: remove core extensions file altogether
