--- lib/onetime/core_ext.rb
@@ -1,30 +1,0 @@
-# lib/onetime/core_ext.rb
-
-
-
-
-# Fix for Mustache 1.1.1 generator compatibility. There is
-# no Mustache::VERSION to check the version. The most
-# recent release was in 2015 so I think we'll be okay.
-if defined?(Mustache)
-  require 'mustache'
-
-  module MustacheGeneratorFix
-    private
-
-    def compile!(exp)
-      case exp.first
-      when :multi
-        exp[1..].reduce(+'') { |sum, e| sum << compile!(e) }
-      when :static
-        str(exp[1])
-      when :mustache
-        send("on_#{exp[1]}", *exp[2..])
-      else
-        raise "Unhandled exp: #{exp.first}"
-      end
-    end
-  end
-
-  Mustache::Generator.prepend(MustacheGeneratorFix)
-end
